services/analytics/requirements.txt
numpy==1.24.3
pandas==2.0.3
scipy==1.10.1
statsmodels==0.14.0
scikit-learn==1.3.0
ruptures==1.1.7
pymannkendall==1.4.2
influxdb-client==1.36.1
psycopg2-binary==2.9.6
sqlalchemy==2.0.19
kafka-python==2.0.2
prometheus-client==0.17.1
python-json-logger==2.0.7
aiohttp==3.8.5
pyyaml==6.0.1
